db without logs

+++ xmlconnectorzos/src/main/java/by/iba/xmlreport/services/status/JobStatusValues.java
@@ -2,5 +2,5 @@ package by.iba.xmlreport.services.status;
 
 public interface JobStatusValues {
     void updateStatus();
-    void updateStatusById(int id);
+    void updateStatusById(long id);
 }
